+++ java-client/src/it/scala/IntegrationTest.scala
@@ -31,6 +31,4 @@ object IntegrationTest {
       }
     }
   }
-
-  def randomString() = UUID.randomUUID.toString
 }